+++ tests/test_operator/test_parse_numeric.py
@@ -21,9 +21,48 @@ def test_parse_add() -> None:
     assert_that(result).is_equal_to(expected_result)
 
 
+def test_parse_divide() -> None:
+    # given
+    sub_dir_path = EXAMPLES_DIR / "divide"
+    file_path = sub_dir_path / "main.bicep"
+    expected_result = json.loads((sub_dir_path / "result.json").read_text())
+
+    # when
+    result = BicepParser(file_path).json()
+
+    # then
+    assert_that(result).is_equal_to(expected_result)
+
+
 def test_parse_minus() -> None:
     # given
     sub_dir_path = EXAMPLES_DIR / "minus"
+    file_path = sub_dir_path / "main.bicep"
+    expected_result = json.loads((sub_dir_path / "result.json").read_text())
+
+    # when
+    result = BicepParser(file_path).json()
+
+    # then
+    assert_that(result).is_equal_to(expected_result)
+
+
+def test_parse_modulo() -> None:
+    # given
+    sub_dir_path = EXAMPLES_DIR / "modulo"
+    file_path = sub_dir_path / "main.bicep"
+    expected_result = json.loads((sub_dir_path / "result.json").read_text())
+
+    # when
+    result = BicepParser(file_path).json()
+
+    # then
+    assert_that(result).is_equal_to(expected_result)
+
+
+def test_parse_multiply() -> None:
+    # given
+    sub_dir_path = EXAMPLES_DIR / "multiply"
     file_path = sub_dir_path / "main.bicep"
     expected_result = json.loads((sub_dir_path / "result.json").read_text())